refact: Implementando interfaces
