+++ src/main/java/br/com/fiap/hackathon/reservafacil/service/ClientService.java
@@ -1,18 +1,7 @@
 package br.com.fiap.hackathon.reservafacil.service;
 
 import br.com.fiap.hackathon.reservafacil.model.Client;
-import br.com.fiap.hackathon.reservafacil.repository.ClientRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Service
-@RequiredArgsConstructor
-public class ClientService {
-    private final ClientRepository repository;
-
-    @Transactional
-    public Client obterPorClientId(String clienteId) {
-        return repository.findByClientId(clienteId);
-    }
+public interface ClientService {
+    Client obterPorClientId(String clienteId);
 }
