+++ src/test/java/br/com/fiap/hackathon/reservafacil/controller/BeneficiarioControllerTest.java
@@ -5,11 +5,11 @@ import br.com.fiap.hackathon.reservafacil.exception.beneficiario.BeneficiarioCad
 import br.com.fiap.hackathon.reservafacil.exception.beneficiario.BeneficiarioNaoEncontradoException;
 import br.com.fiap.hackathon.reservafacil.exception.handler.GlobalExceptionHandler;
 import br.com.fiap.hackathon.reservafacil.exception.usuario.UsuarioNaoIguaisException;
-import br.com.fiap.hackathon.reservafacil.mapper.BeneficiarioMapper;
 import br.com.fiap.hackathon.reservafacil.model.Beneficiario;
 import br.com.fiap.hackathon.reservafacil.model.dto.beneficiario.BeneficiarioResponse;
 import br.com.fiap.hackathon.reservafacil.model.dto.beneficiario.CadastrarBeneficiarioRequest;
 import br.com.fiap.hackathon.reservafacil.service.BeneficiarioService;
+import br.com.fiap.hackathon.reservafacil.service.impl.BeneficiarioServiceImpl;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
@@ -36,7 +36,7 @@ public class BeneficiarioControllerTest {
     private static final String BENEFICIARIO_JA_EXISTE = "Beneficiário já cadastrado";
 
     @Mock
-    private BeneficiarioService service;
+    private BeneficiarioServiceImpl service;
 
     @InjectMocks
     private BeneficiarioController controller;
